+++ Editor/Utils/Util.cs
@@ -1028,26 +1028,22 @@ namespace SaintsField.Editor.Utils
 
         private static (string error, object returnValue) InvokeMethodInfo(MethodInfo methodInfo, object defaultValue, SerializedProperty property, MemberInfo memberInfo, object target, IReadOnlyList<object> overrideParams)
         {
-            object[] passParams;
-            if (property == null || memberInfo == null || target == null)
-            {
-                passParams = Array.Empty<object>();
-            }
-            else
-            {
-                (string error, int arrayIndex, object curValue) = GetValue(property, memberInfo, target);
-                if (error != "")
-                {
-                    return (error, defaultValue);
-                }
+            IReadOnlyList<object> baseParams;
 
-                IReadOnlyList<object> baseParams;
-                if (overrideParams != null)
+            if (overrideParams == null)
+            {
+                if (property == null || memberInfo == null || target == null)
                 {
-                    baseParams = overrideParams;
+                    baseParams = Array.Empty<object>();
                 }
                 else
                 {
+                    (string error, int arrayIndex, object curValue) = GetValue(property, memberInfo, target);
+                    if (error != "")
+                    {
+                        return (error, defaultValue);
+                    }
+
                     baseParams = arrayIndex == -1
                         ? new[]
                         {
@@ -1059,18 +1055,21 @@ namespace SaintsField.Editor.Utils
                             arrayIndex,
                         };
                 }
+            }
+            else
+            {
+                baseParams = overrideParams;
+            }
 
-                string paramError;
-                (paramError, passParams) = ReflectUtils.MethodParamsFill(methodInfo.GetParameters(), baseParams);
-                if (paramError != "")
-                {
-                    return (paramError, null);
-                }
+            (string paramError, object[] passParams) = ReflectUtils.MethodParamsFill(methodInfo.GetParameters(), baseParams);
+            if (paramError != "")
+            {
+                return (paramError, null);
+            }
 
 #if SAINTSFIELD_DEBUG && SAINTSFIELD_DEBUG_UTIL_GET_OF
-                   Debug.Log($"#Util# arrayIndex={arrayIndex}, rawValue={rawValue}, curValue={curValue}, fill={string.Join(",", passParams)}");
+            Debug.Log($"#Util# arrayIndex={arrayIndex}, rawValue={rawValue}, curValue={curValue}, fill={string.Join(",", passParams)}");
 #endif
-            }
 
             object genResult;
             try
@@ -1328,7 +1327,48 @@ namespace SaintsField.Editor.Utils
 
             if (errors.Count == 0)
             {
-                return ($"No method/field/property {fieldOrMethod} found", defaultValue);
+                if (target != null)  // search nested target inside type
+                {
+                    Type targetType = target.GetType();
+                    Type accType = targetType;
+                    foreach (string literType in split)
+                    {
+                        accType = accType.GetNestedType(literType, BindingFlags.Public | BindingFlags.NonPublic);
+                        // Debug.Log($"accType={accType} for {literType}");
+                        if (accType == null)
+                        {
+                            break;
+                        }
+                    }
+
+                    if (accType != null)
+                    {
+                        FieldInfo nestedFieldInfo = type.GetField(fieldOrMethod, bindAttr);
+                        if (nestedFieldInfo != null)
+                        {
+                            object genResult;
+                            try
+                            {
+                                genResult = nestedFieldInfo.GetValue(null);
+                            }
+                            catch (Exception e)
+                            {
+                                // _error = e.Message;
+#if SAINTSFIELD_DEBUG
+                                Debug.LogException(e);
+#endif
+                                return (e.Message, defaultValue);
+                            }
+
+                            return ConvertTo(genResult, defaultValue);
+                        }
+                    }
+
+                }
+#if SAINTSFIELD_DEBUG
+                Debug.LogWarning($"No method/field/property {fieldOrMethod} found for {string.Join(".", split)}");
+#endif
+                return ($"No method/field/property {fieldOrMethod} found for {string.Join(".", split)}", defaultValue);
             }
 
             string finalError = string.Join("\n", errors);
@@ -1904,9 +1944,7 @@ namespace SaintsField.Editor.Utils
 
                 if(!foundResult)
                 {
-                    (string error, object getResult) = conditionStringTarget.Contains(".")
-                        ? AccGetOf<object>(conditionStringTarget, null, property, target, null)
-                        : FlatGetOf<object>(conditionStringTarget, null, property, info, target, null);
+                    (string error, object getResult) = GetOf<object>(conditionStringTarget, null, property, info, target, null);
 
                     if (error != "")
                     {